(#148) Moved JsonSchema class into separate package according to the xml one
